requirements.txt
aiohttp==3.13.3
pydantic>=2.0
python-dotenv==1.2.1
pyyaml>=6.0
qq-adapter-protocol